requirements.txt
tqdm
numpy==1.19.2
tensorflow-gpu==1.15.0
